Refactor server.py to move config classes into base/config.py. (#1854)
