+++ tests/services/api_tests/inventory_test.py
@@ -16,19 +16,19 @@
 
 import time
 import unittest
-from tests.services.inventory import gcp_api_mocks
 from tests.services.api_tests.api_tester import ApiTestRunner
+from tests.services.inventory import gcp_api_mocks
 from tests.services.util.db import create_test_engine
 from tests.services.util.mock import MockServerConfig
 from tests.unittest_utils import ForsetiTestCase
 from google.cloud.forseti.common.util.threadpool import ThreadPool
 from google.cloud.forseti.services import db
+from google.cloud.forseti.services.base.config import InventoryConfig
 from google.cloud.forseti.services.client import ClientComposition
 from google.cloud.forseti.services.dao import ModelManager
 from google.cloud.forseti.services.explain.service import GrpcExplainerFactory
 from google.cloud.forseti.services.inventory.service import GrpcInventoryFactory
 from google.cloud.forseti.services.inventory.storage import Storage
-from google.cloud.forseti.services.server import InventoryConfig
 
 
 class TestServiceConfig(MockServerConfig):
